+++ CHANGES.rst
@@ -40,6 +40,7 @@ Version 3.2.0
     setting or deleting ``response.content_range``.
 -   Modifying ``Response.mimetype_params`` when ``mimetype`` is not set does
     nothing instead of setting an invalid header.
+-   ``WWWAuthenticate.type`` is empty by default instead of ``basic``.
 -   The ``Request.parameter_storage_class``, ``dict_storage_class`` and
     ``list_storage_class`` attributes, and the ``cls`` parameter to
     ``parse_cookie``, ``parse_form_data``, and ``FormDataParser``, are